cliqz: initialize react context on application start
