+++ mozilla-release/mobile/android/base/java/org/mozilla/gecko/GeckoApplication.java
@@ -33,6 +33,7 @@ import android.util.Log;
 import android.util.SparseArray;
 
 import com.cliqz.ABManager;
+import com.cliqz.react.SearchBackground;
 import com.squareup.leakcanary.LeakCanary;
 import com.squareup.leakcanary.RefWatcher;
 
@@ -407,6 +408,7 @@ public class GeckoApplication extends Application
 
         PreferenceManager.init(getApplicationContext());
         ABManager.init(getApplicationContext());
+        SearchBackground.initialize(this);
         /* Cliqz end */
 
         super.onCreate();
